Move ReplacedBy to org.gradle.api.work
